+++ packages/telescope-posts/lib/parameters.js
@@ -101,7 +101,8 @@ Telescope.callbacks.add("postsParameters", limitPosts);
 // hide future scheduled posts unless "showFuture" is set to true or postedAt is already defined
 function hideFuturePosts (parameters, terms) {
 
-  var now = new Date();
+  // var now = new Date();
+  var inOneHour = moment().add(1, "hour").toDate();
 
   if (!parameters.showFuture) {
 
@@ -109,21 +110,21 @@ function hideFuturePosts (parameters, terms) {
     
       if (!!parameters.find.postedAt.$lt) {
 
-        // if postedAt.$lt is defined, use it or current date, whichever is earlier in time
+        // if postedAt.$lt is defined, use it or current date plus one hour, whichever is earlier in time
         var lt = parameters.find.postedAt.$lt;
-        parameters.find.postedAt.$lt = lt < now ? lt : now;
+        parameters.find.postedAt.$lt = lt < inOneHour ? lt : inOneHour;
       
       } else {
 
-        // if postedAt.$lt doesn't exist, use current date
-       parameters.find.postedAt.$lt = now;
+        // if postedAt.$lt doesn't exist, use current date plus one hour
+       parameters.find.postedAt.$lt = inOneHour;
 
       }
 
     } else {
 
-      // if postedAt doesn't exist at all, set it to {$lt: now}
-      parameters.find.postedAt = { $lt: now };
+      // if postedAt doesn't exist at all, set it to {$lt: now plus one hour}
+      parameters.find.postedAt = { $lt: inOneHour };
 
     }